requirements.txt
anyjson==0.3.3
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
django-admin-generator==2.1.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-utils2==2.5.0
djangorestframework==3.9.0
freeze==1.0.10
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
openapi-codec==1.3.2
python-utils==2.3.0
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
